nexdefend-ai/requirements.txt
Flask
psycopg2-binary
numpy
scikit-learn
python-nmap
requests
